# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tongqian Jianzhu AI API
  version: 0.1.0
  description: Shared OpenAPI contract for the Tongqian Jianzhu AI platform.
  license:
    name: Proprietary
    url: https://tongqian.local/license
servers:
  - url: /api/v1
    description: Default API base path
...
paths:
...
  /owner-risk/profiles:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskProfiles
      summary: List owner risk profiles
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of owner risk profiles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskProfileListResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Owner Risk Radar
      operationId: createOwnerRiskProfile
      summary: Create a new owner risk profile
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
        - name: x-user-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOwnerRiskProfileRequest'
      responses:
        '200':
          description: Owner risk profile created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskProfileResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner-risk/profiles/{profileId}:
    get:
      tags:
        - Owner Risk Radar
      operationId: getOwnerRiskProfile
      summary: Get owner risk profile by ID
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Owner risk profile details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskProfileResponse'
        '404':
          description: Owner risk profile not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner-risk/cards:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskCards
      summary: List owner risk cards for the tenant
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of owner risk cards.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskCardListResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner-risk/cards/{cardId}/unlock:
    post:
      tags:
        - Owner Risk Radar
      operationId: unlockOwnerRiskCard
      summary: Unlock owner risk card details
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
        - name: x-user-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Card unlocked successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskCardResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner-risk/reports/{id}:
    get:
      tags:
        - Owner Risk Radar
      operationId: getOwnerRiskReport
      summary: Get owner risk report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskReportResponse'
  /owner-risk/profiles/{profileId}/guarantees:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskGuarantees
      summary: List guarantee records for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of guarantee records.
  /owner-risk/profiles/{profileId}/mixing:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskMixingRecords
      summary: List mixing records for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of mixing records.
  /owner-risk/counterparties:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskCounterparties
      summary: List counterparty watchlist for the tenant
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of counterparty records.
  /owner-risk/profiles/{profileId}/receivables:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskReceivables
      summary: List receivable records for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of receivable records.
  /owner-risk/unlock-logs:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskUnlockLogs
      summary: List unlock logs for the tenant
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of unlock logs.
  /owner-risk/generation-logs:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskGenerationLogs
      summary: List generation logs for the tenant
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of generation logs.
  /owner-risk/profiles/{profileId}/reports:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskReports
      summary: List owner risk reports for a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of reports.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskReportListResponse'
    post:
      tags:
        - Owner Risk Radar
      operationId: createOwnerRiskReport
      summary: Create owner risk report
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
        - name: x-user-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOwnerRiskReportRequest'
      responses:
        '200':
          description: Report created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskReportResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner-risk/profiles/{profileId}/analyze:
    post:
      tags:
        - Owner Risk Radar
      operationId: generateOwnerRiskAnalysis
      summary: Generate owner risk analysis
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
        - name: x-user-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateAnalysisRequest'
      responses:
        '200':
          description: Analysis started successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner-risk/reviews:
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskReviewRequests
      summary: List owner risk review requests
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of review requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskReviewRequestListResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Owner Risk Radar
      operationId: createOwnerRiskReviewRequest
      summary: Submit owner risk review request
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
        - name: x-user-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Review request submitted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskReviewRequestResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Owner Risk Radar
      operationId: listOwnerRiskReviewRequests
      summary: List owner risk review requests
      parameters:
        - name: x-tenant-id
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of review requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerRiskReviewRequestListResponse'
        '4XX':
          description: Request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: BAD_REQUEST
        data:
          type: 'null'
        message:
          type: string
          example: Request failed
        traceId:
          type: string
          example: trace_01JZ0000000000000000000000
    CreateOwnerRiskProfileRequest:
      type: object
      additionalProperties: false
      required:
        - ownerName
      properties:
        ownerName:
          type: string
        idCardMasked:
          type: string
        creditCode:
          type: string
    OwnerRiskProfileView:
      type: object
      additionalProperties: false
      required:
        - id
        - tenantId
        - userId
        - ownerName
        - overallRiskLevel
        - riskScore
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        ownerName:
          type: string
        idCardMasked:
          type: string
        creditCode:
          type: string
        overallRiskLevel:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        guaranteeRiskLevel:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        mixingRiskLevel:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        counterpartyRiskLevel:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        receivableRiskLevel:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        riskScore:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OwnerRiskProfileResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          $ref: '#/components/schemas/OwnerRiskProfileView'
        message:
          type: string
          example: Profile created
        traceId:
          type: string
    OwnerRiskProfileListResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          type: array
          items:
            $ref: '#/components/schemas/OwnerRiskProfileView'
        message:
          type: string
          example: Profiles listed
        traceId:
          type: string
    OwnerRiskCardView:
      type: object
      additionalProperties: false
      required:
        - id
        - profileId
        - tenantId
        - cardType
        - riskLevel
        - isUnlocked
        - unlockCredits
        - createdAt
      properties:
        id:
          type: string
        profileId:
          type: string
        tenantId:
          type: string
        cardType:
          type: string
          enum:
            - guarantee
            - mixing
            - counterparty
            - receivable
        riskLevel:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        confidence:
          type: string
          enum:
            - high
            - medium
            - low
        isUnlocked:
          type: boolean
        unlockCredits:
          type: number
        title:
          type: string
        summary:
          type: string
        detail:
          type: object
        createdAt:
          type: string
          format: date-time
        unlockedAt:
          type: string
          format: date-time
    OwnerRiskCardResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          $ref: '#/components/schemas/OwnerRiskCardView'
        message:
          type: string
          example: Card unlocked
        traceId:
          type: string
    OwnerRiskCardListResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          type: array
          items:
            $ref: '#/components/schemas/OwnerRiskCardView'
        message:
          type: string
          example: Cards listed
        traceId:
          type: string
    CreateOwnerRiskReportRequest:
      type: object
      additionalProperties: false
      required:
        - reportType
      properties:
        reportType:
          type: string
          enum:
            - overview
            - guarantee
            - mixing
            - counterparty
            - receivable
...
    OwnerRiskReportResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            reportId:
              type: string
        message:
          type: string
          example: Report created
        traceId:
          type: string
...
    GenerateAnalysisRequest:
      type: object
      additionalProperties: false
      required:
        - analysisType
      properties:
        analysisType:
          type: string
          enum:
            - summary
            - impact_analysis
            - simulation
            - report
    AnalysisResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          type: object
          properties:
            taskId:
              type: string
        message:
          type: string
          example: Analysis started
        traceId:
          type: string
    CreateReviewRequest:
      type: object
      additionalProperties: false
      required:
        - profileId
        - reviewType
      properties:
        profileId:
          type: string
        reviewType:
          type: string
          enum:
            - human_review
            - expert_consulting
    OwnerRiskReviewRequestView:
      type: object
      additionalProperties: false
      required:
        - id
        - profileId
        - tenantId
        - userId
        - reviewType
        - status
        - createdAt
      properties:
        id:
          type: string
        profileId:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        reviewType:
          type: string
        status:
          type: string
          enum:
            - pending
            - reviewing
            - completed
            - rejected
        assignedTo:
          type: string
        reviewNotes:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    OwnerRiskReviewRequestResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          $ref: '#/components/schemas/OwnerRiskReviewRequestView'
        message:
          type: string
          example: Review submitted
        traceId:
          type: string
    OwnerRiskReviewRequestListResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - data
        - message
        - traceId
      properties:
        code:
          type: string
          example: OK
        data:
          type: array
          items:
            $ref: '#/components/schemas/OwnerRiskReviewRequestView'
        message:
          type: string
          example: Reviews listed
        traceId:
          type: string
...